Move unit test to DataSourceV2RelationSuite in catalyst module
